refactor: 불필요한 쿼리 제거

+++ src/main/java/com/depromeet/buzz/comment/repository/CommentRepositoryCustom.java
@@ -9,5 +9,5 @@ import org.springframework.data.domain.Pageable;
 public interface CommentRepositoryCustom {
     Page<Comment> findCommentsByPostId(Long postId, Pageable pageable);
 
-    Map<Comment, Long> findPopularCommentsByPostId(Long postId);
+    Map<Comment, Integer> findPopularCommentsByPostId(Long postId);
 }
